requirements.txt
pydantic
python-dotenv
langchain
langchain-groq
langchain-core
langchain-community
langchain_huggingface
langgraph

defusedxml 
pydantic_extra_types
pycountry
pytesseract
pymupdf

libmagic

langchain_milvus 
rank_bm25

scrapingant-client